use -file for directory, too
Currently, -file and -dir option exists seperatively.  However, it does
not have to be so and a little annoying to use the interface in some
case.  For the reason, this commit unifies those options into one.  With
this commit, user can pass path to file or directory via -file option.
Program will check whether the path is to file or directory and do works
appropriately.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>